Split main into multiple packages
