+++ pkg/cgroup/cgroup.go
@@ -1,0 +1,7 @@
+package cgroup
+
+import "k8s.io/apimachinery/pkg/types"
+
+type CgroupHandler interface {
+	WriteCPUMax(podUID types.UID, containerID string, cpuMaxValue uint64) error
+}
